Add kwm and khd

+++ install.sh
@@ -13,14 +13,21 @@ done
 
 echo "installing nvimrc"
 mkdir -p ~/.config/nvim
-ln -s $current_dir/nvimrc ~/.config/nvim/init.vim
+if [ ! -f ~/.config/nvim/init.vim ]; then
+  ln -s $current_dir/nvimrc ~/.config/nvim/init.vim
+fi
 
 echo "installing vim-plug"
-curl -fLo ~/.nvim/autoload/plug.vim --create-dirs \
-    https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+if [ ! -f ~/.nvim/autoload/plug.vim ]; then
+  curl -fLo ~/.nvim/autoload/plug.vim --create-dirs \
+      https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+fi
 
 echo "installing zshrc"
-git clone git@github.com:zsh-users/antigen.git ~/.antigen-repo
+if [ ! -d ~/.antigen-repo ]; then
+  git clone git@github.com:zsh-users/antigen.git ~/.antigen-repo
+fi
+
 if [ ! -f ~/.zshrc ]; then
   cp zshrc ~/.zshrc
 fi
@@ -30,6 +37,15 @@ mkdir -p ~/.emacs.d
 if [ ! -f ~/.emacs.d/init.el ]; then
   cp init.el ~/.emacs.d/init.el
   touch ~/.emacs.d/custom.el
+fi
+
+echo "installing kwm"
+mkdir -p ~/.kwm
+if [ ! -f ~/.kwm/kwmrc ]; then
+  ln -s $current_dir/kwmrc ~/.kwm/kwmrc
+fi
+if [ ! -f ~/.khdrc ]; then
+  ln -s $current_dir/khdrc ~/.khdrc
 fi
 
 echo "complete, don't forget to update the .zshrc"
